requirements.txt
asgiref==3.10.0
blinker==1.9.0
click==8.3.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3

pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
runserver==1.0.4
sqlparse==0.5.3
Werkzeug==3.1.3
